+++ packages/widgetbook_generator/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - **FEAT**: Support nullable args. ([#1675](https://github.com/widgetbook/widgetbook/pull/1675))
+- **REFACTOR**: Initialize default args in field initializers, instead of `StoryArgs` constructor. ([#1681](https://github.com/widgetbook/widgetbook/pull/1681))
 
 ## 4.0.0-alpha.3